Added try and catch to deleteCategory

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -75,23 +75,34 @@ public class CategoriesController {
         }
 
     }
-        // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
-        // add annotation to ensure that only an ADMIN can call this function
+    // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
+    // add annotation to ensure that only an ADMIN can call this function
 
-        @PutMapping("{categoryId}")
-        @PreAuthorize("hasRole('ROLE_ADMIN')")
-        public void updateCategory ( @PathVariable int categoryId, @RequestBody Category category){
-            categoryDao.update(categoryId, category);
-        }
+    @PutMapping("{categoryId}")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    public void updateCategory(@PathVariable int categoryId, @RequestBody Category category) {
+        categoryDao.update(categoryId, category);
+    }
 
 
-        // add annotation to call this method for a DELETE action - the url path must include the categoryId
-        // add annotation to ensure that only an ADMIN can call this function
+    // add annotation to call this method for a DELETE action - the url path must include the categoryId
+    // add annotation to ensure that only an ADMIN can call this function
 
-        @DeleteMapping("{categoryId}")
-        @PreAuthorize("hasRole('ROLE_ADMIN')")
-        public void deleteCategory ( @PathVariable int categoryId){
+    @DeleteMapping("{categoryId}")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    public void deleteCategory(@PathVariable int categoryId) {
+
+        try {
+            var category = categoryDao.getById(categoryId);
+
+            if (category == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
             categoryDao.delete(categoryId);
-            // delete the category by id
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
         }
     }
+}
+
+            // delete the category by id
